Move into xarray submodule
